refactor tests, extract loading/parsing functions into a dedicated class
